Finished designing and running the solution code
The code was ran a few times, on the same set of dominoes using
different SA parameters. Two different solutions were found, denoted by
a suffixed "1" or "2".

"domino_solver.py" -- contains classes and fuctions for solving the
Domino puzzle levels from the Android app 'Mind Games'
"tester2.py" -- imports and runs 'domino_solver' to actually solve the
puzzles
"Solution_Output_.txt" -- what was printed to the screen when the code
had finished running
"Successful_Level2_.txt" -- a graph of the progress of the SA algorithm
"best_list_.txt" -- a text file containing the cost of the best domino
configuration found at that point in the algorithm
"cost_list_.txt" -- a text file containing the cost of the domino
configuration at every step of the algorithm